+++ civis/tests/test_io.py
@@ -186,7 +186,9 @@ def test_civis_file_to_table_table_exists(m_run_cleaning,
     mock_civis.get_database_id.return_value = 42
     mock_civis.default_credential = 713
 
-    mock_civis.get_table_id.return_value = 42
+    mock_civis.databases.get_schemas_tables.return_value = Response({
+        'name': 'table1'
+    })
     m_process_cleaning_results.return_value = (
         [mock_cleaned_file_id],
         True,  # headers
@@ -280,7 +282,7 @@ def test_civis_file_to_table_table_doesnt_exist(m_run_cleaning,
     mock_civis.get_database_id.return_value = 42
     mock_civis.default_credential = 713
 
-    mock_civis.get_table_id.side_effect = ValueError('no table')
+    mock_civis.databases.get_schemas_tables.side_effect = MockAPIError(404)
     mock_columns = [{'name': 'foo', 'sql_type': 'INTEGER'}]
     m_process_cleaning_results.return_value = (
         [mock_cleaned_file_id],
@@ -375,7 +377,7 @@ def test_civis_file_to_table_table_doesnt_exist_all_sql_types_missing(
     mock_civis.imports.post_files_csv.return_value.id = mock_import_id
     mock_civis.get_database_id.return_value = 42
     mock_civis.default_credential = 713
-    mock_civis.get_table_id.side_effect = ValueError('no table')
+    mock_civis.databases.get_schemas_tables.side_effect = MockAPIError(404)
     table_columns = [{'name': 'a', 'sql_type': ''},
                      {'name': 'b', 'sql_type': ''}]
     detected_columns = [{'name': 'a', 'sql_type': 'INTEGER'},
@@ -474,7 +476,7 @@ def test_civis_file_to_table_table_does_not_exist_some_sql_types_missing(
     mock_civis.imports.post_files_csv.return_value.id = mock_import_id
     mock_civis.get_database_id.return_value = 42
     mock_civis.default_credential = 713
-    mock_civis.get_table_id.side_effect = ValueError('no table')
+    mock_civis.databases.get_schemas_tables.side_effect = MockAPIError(404)
     table_columns = [{'name': 'a', 'sql_type': 'INT'},
                      {'name': 'b', 'sql_type': ''}]
 
@@ -506,7 +508,7 @@ def test_civis_file_to_table_table_columns_keys_misspelled(
     mock_civis.imports.post_files_csv.return_value.id = mock_import_id
     mock_civis.get_database_id.return_value = 42
     mock_civis.default_credential = 713
-    mock_civis.get_table_id.side_effect = ValueError('no table')
+    mock_civis.databases.get_schemas_tables.side_effect = MockAPIError(404)
     table_columns = [{'name': 'a', 'sqlType': 'INT'},
                      {'name': 'b', 'bad_type': ''}]
 
@@ -544,7 +546,7 @@ def test_civis_file_to_table_table_doesnt_exist_provide_table_columns(
         mock_civis.imports.post_files_csv.return_value.id = mock_import_id
         mock_civis.get_database_id.return_value = 42
         mock_civis.default_credential = 713
-        mock_civis.get_table_id.side_effect = ValueError('no table')
+        mock_civis.databases.get_schemas_tables.side_effect = MockAPIError(404)
         table_columns = [{'name': 'foo', 'sql_type': 'INTEGER'},
                          {'name': 'bar', 'sql_type': 'VARCHAR(42)'}]
         m_process_cleaning_results.return_value = (
@@ -654,7 +656,7 @@ def test_civis_file_to_table_multi_file(
     mock_civis.get_database_id.return_value = 42
     mock_civis.default_credential = 713
 
-    mock_civis.get_table_id.side_effect = ValueError('no table')
+    mock_civis.databases.get_schemas_tables.side_effect = MockAPIError(404)
     mock_columns = [{'name': 'foo', 'sql_type': 'INTEGER'}]
     m_process_cleaning_results.return_value = (
         mock_cleaned_file_ids,